backend/requirements.txt
fastapi[all]
uvicorn[standard] 
python-multipart 
transformers 
pillow 
requests